Move null view sanitization into the Stitcher (QTF producer)
Run it on the hand-built QTF output instead of in the shared reduce sinks,
so only the small K-row batch is scanned and the generic path is untouched.
