+++ sandbox/plugins/analytics-backend-datafusion/src/main/java/org/opensearch/be/datafusion/DatafusionReduceSink.java
@@ -222,7 +222,6 @@ public class DatafusionReduceSink extends AbstractDatafusionReduceSink implement
                     + batch.getSchema()
             );
         }
-        ViewVectorSanitizer.sanitize(batch);
         ArrowArray array = ArrowArray.allocateNew(alloc);
         ArrowSchema arrowSchema = ArrowSchema.allocateNew(alloc);
         try {
